little bug in admin index and make redirects uniform
admin index would throw an error if the first thing you did after getting to admin cp from the installer was switch to dark mode.

Also changed the redirects in the installer to all be the same (removed the installer.php from the half that had it)

+++ modules/Core/pages/admin/index.php
@@ -18,10 +18,12 @@ if($user->isLoggedIn()){
 	} else {
 		// Try to delete installer
 		if(isset($_GET['from']) && $_GET['from'] == 'install'){
-			try {
-				unlink('install.php');
-			} catch(Exception $e){
-				die('Unable to delete installer automatically, please remove <strong>install.php</strong> from your website before continuing.');
+			if(is_file('install.php')){
+				try {
+					unlink('install.php');
+				} catch(Exception $e){
+					die('Unable to delete installer automatically, please remove <strong>install.php</strong> from your website before continuing.');
+				}
 			}
 		}
